move `PipeSystem` into `combinator.rs`
