--- crates/bevy_ecs/src/system/system_piping.rs
@@ -1,65 +1,0 @@
-use crate::system::System;
-
-use super::{CombinatorSystem, Combine};
-
-/// A [`System`] created by piping the output of the first system into the input of the second.
-///
-/// This can be repeated indefinitely, but system pipes cannot branch: the output is consumed by the receiving system.
-///
-/// Given two systems `A` and `B`, A may be piped into `B` as `A.pipe(B)` if the output type of `A` is
-/// equal to the input type of `B`.
-///
-/// Note that for [`FunctionSystem`](crate::system::FunctionSystem)s the output is the return value
-/// of the function and the input is the first [`SystemParam`](crate::system::SystemParam) if it is
-/// tagged with [`In`](crate::system::In) or `()` if the function has no designated input parameter.
-///
-/// # Examples
-///
-/// ```
-/// use std::num::ParseIntError;
-///
-/// use bevy_ecs::prelude::*;
-///
-/// fn main() {
-///     let mut world = World::default();
-///     world.insert_resource(Message("42".to_string()));
-///
-///     // pipe the `parse_message_system`'s output into the `filter_system`s input
-///     let mut piped_system = parse_message_system.pipe(filter_system);
-///     piped_system.initialize(&mut world);
-///     assert_eq!(piped_system.run((), &mut world), Some(42));
-/// }
-///
-/// #[derive(Resource)]
-/// struct Message(String);
-///
-/// fn parse_message_system(message: Res<Message>) -> Result<usize, ParseIntError> {
-///     message.0.parse::<usize>()
-/// }
-///
-/// fn filter_system(In(result): In<Result<usize, ParseIntError>>) -> Option<usize> {
-///     result.ok().filter(|&n| n < 100)
-/// }
-/// ```
-pub type PipeSystem<SystemA, SystemB> = CombinatorSystem<Pipe, SystemA, SystemB>;
-
-#[doc(hidden)]
-pub struct Pipe;
-
-impl<A, B> Combine<A, B> for Pipe
-where
-    A: System,
-    B: System<In = A::Out>,
-{
-    type In = A::In;
-    type Out = B::Out;
-
-    fn combine(
-        input: Self::In,
-        a: impl FnOnce(<A as System>::In) -> <A as System>::Out,
-        b: impl FnOnce(<B as System>::In) -> <B as System>::Out,
-    ) -> Self::Out {
-        let value = a(input);
-        b(value)
-    }
-}
